+++ resources/views/people/export.blade.php
@@ -3,15 +3,18 @@
             <tr>
                 <th>@lang('people.family_name')</th>
                 <th>@lang('people.name')</th>
-                <th>@lang('people.police_no')</th>
-                <th>@lang('people.case_no')</th>
-                <th>@lang('people.med_no')</th>
-                <th>@lang('people.reg_no')</th>
-                <th>@lang('people.sec_card_no')</th>
-                <th>@lang('people.temp_no')</th>
+                <th>@lang('people.date_of_birth')</th>
+                <th>@lang('people.age')</th>
                 <th>@lang('people.nationality')</th>
+                <th>@lang('people.police_number')</th>
+                <th>@lang('people.case_number')</th>
+                <th>@lang('people.medical_number')</th>
+                <th>@lang('people.registration_number')</th>
+                <th>@lang('people.section_card_number')</th>
+                <th>@lang('people.temporary_number')</th>
                 <th>@lang('people.languages')</th>
                 <th>@lang('people.skills')</th>
+                <th>@lang('app.registered')</th>
                 <th>@lang('people.remarks')</th>
             </tr>
         </thead>
@@ -20,15 +23,18 @@
                 <tr>
                     <td>{{ $person->family_name }}</td>
                     <td>{{ $person->name }}</td>
+                    <td>{{ $person->date_of_birth }}</td>
+                    <td>{{ $person->age }}</td>
+                    <td>{{ $person->nationality }}</td>
                     <td>{{ $person->police_no }}</td>
                     <td>{{ $person->case_no }}</td>
                     <td>{{ $person->medical_no }}</td>
                     <td>{{ $person->registration_no }}</td>
                     <td>{{ $person->section_card_no }}</td>
                     <td>{{ $person->temp_no }}</td>
-                    <td>{{ $person->nationality }}</td>
                     <td>{{ $person->languages }}</td>
                     <td>{{ $person->skills }}</td>
+                    <td>{{ $person->created_at->toDateString() }}</td>
                     <td>{{ $person->remarks }}</td>
                 </tr>
             @endforeach